Snake code
This is the code I used when I made snake in unity. There is two files that one of them is code for the food to spawn in a random place when the snake eats the food, and last file is code making the character snake to move and eat food.